Added initial preferences activity tests, mild refactor
