--- project/app/src/androidTest/java/org/owntracks/android/ui/map/Preferences.kt
@@ -1,32 +1,0 @@
-package org.owntracks.android.ui.map
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.filters.LargeTest
-import com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed
-import com.schibsted.spain.barista.rule.BaristaRule
-import com.schibsted.spain.barista.rule.flaky.AllowFlaky
-import org.junit.Before
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.owntracks.android.R
-import org.owntracks.android.ui.preferences.PreferencesActivity
-
-@LargeTest
-@RunWith(AndroidJUnit4::class)
-class Preferences {
-    @get:Rule
-    var baristaRule = BaristaRule.create(PreferencesActivity::class.java)
-
-    @Before
-    fun setUp() {
-        baristaRule.launchActivity()
-    }
-
-    @Test
-    @AllowFlaky(attempts = 1)
-    fun infoSectionsLaunchExternalActivities() {
-        assertDisplayed(R.string.preferencesInfo)
-
-    }
-}
